requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.31
json5==0.9